+++ kinto/core/testing.py
@@ -111,22 +111,23 @@ class BaseWebTest:
     entry_point = None
     """Main application entry"""
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.app = self.make_app()
-        self.storage = self.app.app.registry.storage
-        self.cache = self.app.app.registry.cache
-        self.permission = self.app.app.registry.permission
+    headers = {
+        'Content-Type': 'application/json'
+    }
 
-        self.storage.initialize_schema()
-        self.permission.initialize_schema()
-        self.cache.initialize_schema()
+    @classmethod
+    def setUpClass(cls):
+        cls.app = cls.make_app()
+        cls.storage = cls.app.app.registry.storage
+        cls.cache = cls.app.app.registry.cache
+        cls.permission = cls.app.app.registry.permission
 
-        self.headers = {
-            'Content-Type': 'application/json'
-        }
+        cls.storage.initialize_schema()
+        cls.permission.initialize_schema()
+        cls.cache.initialize_schema()
 
-    def make_app(self, settings=None, config=None):
+    @classmethod
+    def make_app(cls, settings=None, config=None):
         """Instantiate the application and setup requests to use the api
         prefix.
 
@@ -134,19 +135,17 @@ class BaseWebTest:
         :param pyramid.config.Configurator config: already initialized config
         :returns: webtest application instance
         """
-        settings = self.get_app_settings(extras=settings)
+        settings = cls.get_app_settings(extras=settings)
 
-        try:
-            main = self.entry_point.__func__
-        except AttributeError:  # pragma: no cover
-            main = self.entry_point.im_func
+        main = cls.entry_point
 
         wsgi_app = main({}, config=config, **settings)
         app = webtest.TestApp(wsgi_app)
-        app.RequestClass = get_request_class(self.api_prefix)
+        app.RequestClass = get_request_class(cls.api_prefix)
         return app
 
-    def get_app_settings(self, extras=None):
+    @classmethod
+    def get_app_settings(cls, extras=None):
         """Application settings to be used. Override to tweak default settings
         for the tests.