+++ tests/core/resource/test_events.py
@@ -30,24 +30,27 @@ def notif_broken(app, event_cls):
 class BaseEventTest(BaseWebTest):
 
     subscribed = tuple()
+    events = []
 
     def setUp(self):
         super().setUp()
-        self.events = []
+        del self.events[:]
         self.body = {'data': {'name': 'de Paris'}}
 
     def tearDown(self):
-        self.events = []
+        del self.events[:]
         super().tearDown()
 
-    def listener(self, event):
-        self.events.append(event)
+    @classmethod
+    def listener(cls, event):
+        cls.events.append(event)
 
-    def make_app(self, settings=None, config=None):
-        settings = self.get_app_settings(settings)
+    @classmethod
+    def make_app(cls, settings=None, config=None):
+        settings = cls.get_app_settings(settings)
         config = Configurator(settings=settings)
-        for event_cls in self.subscribed:
-            config.add_subscriber(self.listener, event_cls)
+        for event_cls in cls.subscribed:
+            config.add_subscriber(cls.listener, event_cls)
         config.commit()
         return super().make_app(settings=settings, config=config)
 
@@ -76,7 +79,7 @@ class ResourceReadTest(BaseEventTest, unittest.TestCase):
         resp = self.app.post_json(self.collection_url, self.body,
                                   headers=self.headers, status=201)
         record = resp.json['data']
-        body = dict(self.body)
+        body = {**self.body}
         body['data']['id'] = record['id']
 
         # a second post with the same record id
@@ -105,7 +108,7 @@ class ResourceChangedTest(BaseEventTest, unittest.TestCase):
                          ACTIONS.CREATE.value)
 
     def test_put_sends_create_action(self):
-        body = dict(self.body)
+        body = {**self.body}
         body['data']['id'] = record_id = str(uuid.uuid4())
         record_url = self.get_item_url(record_id)
         self.app.put_json(record_url, body,
@@ -115,11 +118,12 @@ class ResourceChangedTest(BaseEventTest, unittest.TestCase):
                          ACTIONS.CREATE.value)
 
     def test_not_triggered_on_failed_put(self):
-        record_id = str(uuid.uuid4())
+        body = {**self.body}
+        body['data']['id'] = record_id = str(uuid.uuid4())
         record_url = self.get_item_url(record_id)
-        self.app.put_json(record_url, self.body, headers=self.headers)
+        self.app.put_json(record_url, body, headers=self.headers)
         headers = {**self.headers, 'If-Match': '"12345"'}
-        self.app.put_json(record_url, self.body, headers=headers, status=412)
+        self.app.put_json(record_url, body, headers=headers, status=412)
         self.assertEqual(len(self.events), 1)
         self.assertEqual(self.events[0].payload['action'],
                          ACTIONS.CREATE.value)
@@ -139,7 +143,7 @@ class ResourceChangedTest(BaseEventTest, unittest.TestCase):
                          ACTIONS.UPDATE.value)
 
     def test_put_sends_update_action_if_record_exists(self):
-        body = dict(self.body)
+        body = {**self.body}
         body['data']['id'] = record_id = str(uuid.uuid4())
         record_url = self.get_item_url(record_id)
         self.app.put_json(record_url, body,
@@ -429,7 +433,9 @@ def load_from_config(config, prefix):
 
 @unittest.skipIf(not statsd.statsd_module, "statsd is not installed.")
 class StatsDTest(BaseWebTest, unittest.TestCase):
-    def get_app_settings(self, *args, **kwargs):
+
+    @classmethod
+    def get_app_settings(cls, *args, **kwargs):
         settings = super().get_app_settings(*args, **kwargs)
         if not statsd.statsd_module:
             return settings