+++ tests/plugins/test_history.py
@@ -27,7 +27,8 @@ class PluginSetup(unittest.TestCase):
 
 class HistoryWebTest(support.BaseWebTest, unittest.TestCase):
 
-    def get_app_settings(self, extras=None):
+    @classmethod
+    def get_app_settings(cls, extras=None):
         settings = super().get_app_settings(extras)
         settings['includes'] = 'kinto.plugins.history'
         return settings
@@ -527,7 +528,8 @@ class BulkTest(HistoryWebTest):
 
 class DefaultBucketTest(HistoryWebTest):
 
-    def get_app_settings(self, extras=None):
+    @classmethod
+    def get_app_settings(cls, extras=None):
         settings = super().get_app_settings(extras)
         settings['includes'] = ('kinto.plugins.default_bucket '
                                 'kinto.plugins.history')
@@ -580,7 +582,8 @@ class DefaultBucketTest(HistoryWebTest):
 
 class PermissionsTest(HistoryWebTest):
 
-    def get_app_settings(self, extras=None):
+    @classmethod
+    def get_app_settings(cls, extras=None):
         settings = super().get_app_settings(extras)
         settings['experimental_permissions_endpoint'] = 'true'
         return settings
@@ -714,7 +717,8 @@ class PermissionsTest(HistoryWebTest):
 
 class ExcludeResourcesTest(HistoryWebTest):
 
-    def get_app_settings(self, extras=None):
+    @classmethod
+    def get_app_settings(cls, extras=None):
         settings = super().get_app_settings(extras)
         settings['history.exclude_resources'] = ('/buckets/a '
                                                  '/buckets/b/collections/a '